+++ CHANGELOG.md
@@ -20,6 +20,8 @@ Also possible, but for this project less relevant, is `Deprecated` for soon-to-b
 
 ## Unreleased
 
+### Deprecated
+* Some physics-related input keys in `Output: Initial_Conditions` were deprecated and duplicated under `Modi: Collider: Initial_Conditions`
 
 ## SMASH-3.1
 Date: 2024-02-26